+++ vice/src/arch/win32/ps-dll.c
@@ -109,7 +109,7 @@ static BYTE parsid_inb(unsigned int addrint)
         winio_inp32fp(addr, &tmp, 1);
         retval = (BYTE)tmp;
     } else {
-        retval = inpout_inp32fp(addr);
+        retval = (BYTE)inpout_inp32fp(addr);
     }
     return retval;
 }
